Add service discovery
Add service discovery via Consul so that no matter how many nodes are
added, they will all be updated. Technically this takes away from it
being decentralized, but the goal of this example isn't to develop a
fully fledged P2P network!

Also added a configuration file, env vars and console parameters to that
it's easier to get multiple nodes up and running